+++ TruRating.TruModule.V2xx.ConsoleRunner/Program.cs
@@ -19,7 +19,6 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-
 using System;
 using TruRating.TruModule.V2xx.ConsoleRunner.Device;
 using TruRating.TruModule.V2xx.ConsoleRunner.Environment;
