+++ TruRating.TruModule.V2xx.ConsoleRunner/UseCase/UseCaseFactory.cs
@@ -19,7 +19,6 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-
 using System.Collections.Generic;
 using TruRating.TruModule.V2xx.ConsoleRunner.Environment;
 using TruRating.TruModule.V2xx.Device;
