+++ TruRating.TruModule.V2xx/Helpers/TaskHelpers.cs
@@ -19,7 +19,6 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-
 using System.Threading;
 
 namespace TruRating.TruModule.V2xx.Helpers
