+++ TruRating.TruModule.V2xx/Network/IWebClientFactory.cs
@@ -19,21 +19,10 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-using System;
-
-namespace TruRating.TruModule.V2xx.Helpers
+namespace TruRating.TruModule.V2xx.Network
 {
-    public static class DateTimeProvider
+    public interface IWebClientFactory
     {
-        private static Func<DateTime> _nowFunc = () => DateTime.UtcNow;
-
-        /// <summary>
-        ///     The current representation of UtcNow
-        /// </summary>
-        public static DateTime UtcNow
-        {
-            get { return _nowFunc(); }
-            set { _nowFunc = () => value; }
-        }
+        IWebClient Create();
     }
 }
