+++ TruRating.TruModule.V2xx/Serialization/ISerializer.cs
@@ -19,7 +19,6 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-
 namespace TruRating.TruModule.V2xx.Serialization
 {
     public interface ISerializer
