+++ TruRating.TruModule.V2xx/Serialization/TruServiceMessageFactory.cs
@@ -19,7 +19,6 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-
 using TruRating.Dto.TruService.V220;
 using TruRating.Dto.TruService.V2xx;
 using TruRating.TruModule.V2xx.Device;
